requirements.txt
# bert-score==0.3.13
# certifi==2025.4.26
# charset-normalizer==3.4.2
# contourpy==1.3.0
# cycler==0.12.1
# filelock==3.18.0
# fonttools==4.58.0
# fsspec==2025.3.2
# hf-xet==1.1.0
# huggingface-hub==0.31.1
# idna==3.10
# importlib_resources==6.5.2
# Jinja2==3.1.6
# kiwisolver==1.4.7
# MarkupSafe==3.0.2
# matplotlib==3.9.4
# mpmath==1.3.0
# networkx==3.2.1
# numpy==2.0.2
# nvidia-cublas-cu12==12.6.4.1
# nvidia-cuda-cupti-cu12==12.6.80
# nvidia-cuda-nvrtc-cu12==12.6.77
# nvidia-cuda-runtime-cu12==12.6.77
# nvidia-cudnn-cu12==9.5.1.17
# nvidia-cufft-cu12==11.3.0.4
# nvidia-cufile-cu12==1.11.1.6
# nvidia-curand-cu12==10.3.7.77
# nvidia-cusolver-cu12==11.7.1.2
# nvidia-cusparse-cu12==12.5.4.2
# nvidia-cusparselt-cu12==0.6.3
# nvidia-nccl-cu12==2.26.2
# nvidia-nvjitlink-cu12==12.6.85
# nvidia-nvtx-cu12==12.6.77
# packaging==25.0
# pandas==2.2.3
# pillow==11.2.1
# pyparsing==3.2.3
# python-dateutil==2.9.0.post0
# pytz==2025.2
# PyYAML==6.0.2
# regex==2024.11.6
# requests==2.32.3
# safetensors==0.5.3
# six==1.17.0
# sympy==1.14.0
# tokenizers==0.21.1
# torch==2.7.0
# tqdm==4.67.1
# transformers==4.51.3
# triton==3.3.0
# typing_extensions==4.13.2
# tzdata==2025.2
# urllib3==2.4.0
# zipp==3.21.0






# torch>=2.0.0
# transformers==4.38.2
# peft==0.7.1
# pillow>=9.5.0
# pandas>=2.0.0
# tqdm>=4.65.0
# gdown>=4.7.1
# bert-score>=0.3.13
# pytz>=2023.3


# Core ML and Transformers
torch>=2.0.0
transformers==4.51.3
peft==0.15.2

# Image and Data Handling
pillow>=9.5.0
pandas>=2.0.0
tqdm>=4.65.0

# Google Drive Download
gdown>=4.7.1

# Evaluation Metric
bert-score>=0.3.13
